# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nomi Core",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/commitments": {
      "post": {
        "summary": "Create Commitment Route",
        "operationId": "create_commitment_route_api_v1_commitments_post",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommitmentInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommitmentOut"
                }
              }
            }
          },
          "400": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "408": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "413": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "415": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Commitments Route",
        "operationId": "list_commitments_route_api_v1_commitments_get",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 30,
              "title": "Limit"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "all",
                "receivable",
                "payable",
                "overdue",
                "due_soon"
              ],
              "type": "string",
              "default": "all",
              "title": "Filter"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 160,
              "default": "",
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_CommitmentListOut_"
                }
              }
            }
          },
          "400": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "408": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "413": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "415": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/commitments/{commitment_id}": {
      "get": {
        "summary": "Get Commitment",
        "operationId": "get_commitment_api_v1_commitments__commitment_id__get",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "commitment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Commitment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommitmentOut"
                }
              }
            }
          },
          "400": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "408": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "413": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "415": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Edit Commitment",
        "operationId": "edit_commitment_api_v1_commitments__commitment_id__patch",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "commitment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Commitment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommitmentEdit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommitmentOut"
                }
              }
            }
          },
          "400": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "408": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "413": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "415": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/commitments/{commitment_id}/cancel": {
      "post": {
        "summary": "Cancel Commitment",
        "operationId": "cancel_commitment_api_v1_commitments__commitment_id__cancel_post",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "commitment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Commitment Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommitmentVersion"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommitmentOut"
                }
              }
            }
          },
          "400": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "408": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "413": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "415": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CommitmentEdit": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "maximum": 2147483646.0,
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "concept": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 160
              },
              {
                "type": "null"
              }
            ],
            "title": "Concept"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version"
        ],
        "title": "CommitmentEdit"
      },
      "CommitmentInput": {
        "properties": {
          "contact_id": {
            "type": "string",
            "format": "uuid",
            "title": "Contact Id"
          },
          "direction": {
            "type": "string",
            "enum": [
              "receivable",
              "payable"
            ],
            "title": "Direction"
          },
          "original_amount_minor": {
            "type": "integer",
            "maximum": 9000000000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Original Amount Minor"
          },
          "currency_code": {
            "type": "string",
            "pattern": "^[A-Z]{3}$",
            "title": "Currency Code"
          },
          "concept": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 160
              },
              {
                "type": "null"
              }
            ],
            "title": "Concept"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "contact_id",
          "direction",
          "original_amount_minor",
          "currency_code"
        ],
        "title": "CommitmentInput"
      },
...
      "CommitmentOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "contact_id": {
            "type": "string",
            "format": "uuid",
            "title": "Contact Id"
          },
          "direction": {
            "type": "string",
            "enum": [
              "receivable",
              "payable"
            ],
            "title": "Direction"
          },
          "original_amount_minor": {
            "type": "integer",
            "title": "Original Amount Minor"
          },
          "balance_minor": {
            "type": "integer",
            "title": "Balance Minor"
          },
          "currency_code": {
            "type": "string",
            "title": "Currency Code"
          },
          "concept": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Concept"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "lifecycle_status": {
            "type": "string",
            "enum": [
              "open",
              "paid",
              "cancelled"
            ],
            "title": "Lifecycle Status"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "payment_state": {
            "type": "string",
            "enum": [
              "pending",
              "partial",
              "paid"
            ],
            "title": "Payment State"
          },
          "timing_state": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "no_due_date",
                  "upcoming",
                  "due_soon",
                  "overdue"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Timing State"
          }
        },
        "type": "object",
        "required": [
          "id",
          "contact_id",
          "direction",
          "original_amount_minor",
          "balance_minor",
          "currency_code",
          "concept",
          "due_date",
          "lifecycle_status",
          "version",
          "payment_state",
          "timing_state"
        ],
        "title": "CommitmentOut"
      },
      "CommitmentVersion": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "maximum": 2147483646.0,
            "minimum": 1.0,
            "title": "Expected Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version"
        ],
        "title": "CommitmentVersion"
      },
...
      "Page_CommitmentListOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CommitmentListOut"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "Page[CommitmentListOut]"
      },
...
      "Problem": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "instance": {
            "type": "string",
            "title": "Instance"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code",
          "request_id"
        ],
        "title": "Problem"
      },
...
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "nomi_session"
      }
    }
  }
}